+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## kedro-neptune 0.0.6
+
+### Features
+- Added `enabled` flag to neptune.yml to maintain that metadata logging into Neptune is enabled ([#20](https://github.com/neptune-ai/kedro-neptune/pull/20))
+
 ## kedro-neptune 0.0.5
 
 ### Fixes